cam: Add Consumer and restructure KMS and Net
Add Consumer ABC which represents a "consumer" for the frames. KMS and
Net-tx are the currently available consumers.

The KMS and Net Consumers should be merged with the already existing net
and kms classes.

The KMS being a buffer allocator and consumer is a bit messy. They
should be split.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>