cam: Apply sensor controls in correct order relative to format setting
Some sensor controls—such as VBLANK, HBLANK, and EXPOSURE have ranges
that depend on the active resolution and are updated by the driver when
a new format is set. To avoid having these values clamped or reset,
ensure they are applied after the format is configured.

Conversely, certain controls like VFLIP and HFLIP must be applied before
setting the format. To support this, introduce a pre-controls dictionary
in the configuration for controls that need to be set prior to the
format update.

Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>